% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisAnnotate.R
\name{basisAnnotate}
\alias{basisAnnotate}
\title{Annotate the bases of scPNMF}
\usage{
basisAnnotate(
  W,
  DE_prop = 0.1,
  dim_use = NULL,
  id_type = "SYMBOL",
  ont = "BP",
  OrgDb = "org.Hs.eg.db",
  pvalueCutoff = 0.1,
  qvalueCutoff = 0.1,
  minGSSize = 20,
  maxGSSize = 100,
  simp = FALSE,
  return_fig = FALSE,
  cat_num = 10,
  word_num = 50
)
}
\arguments{
\item{W}{The weight matrix from scPNMF.}

\item{DE_prop}{The proportion of genes to be treated as "DE" genes for GO analysis.}

\item{dim_use}{Which bases to annotate.}

\item{id_type}{The gene id types. For example, official gene symbol is \code{SYMBOL}. See details in package \code{clusterProfiler}.}

\item{ont}{Which gene ontology to use. One of "BP", "MF" or "CC". The default is "BP".}

\item{OrgDb}{OrgDb to use.}

\item{pvalueCutoff}{P-value cutoff on on enrichment tests to report. Default is 0.1.}

\item{qvalueCutoff}{Q-value cutoff on on enrichment tests to report. Default is 0.1.}

\item{minGSSize}{Minimal size of genes annotated by Ontology term for testing. Default is 20.}

\item{maxGSSize}{Maximal size of genes annotated by Ontology term for testing. Default is 100.}

\item{simp}{If simplifying the GO terms. Default is FALSE.}

\item{return_fig}{If returning the visualization plots. Default is FALSE.}

\item{cat_num}{How many GO terms will be shown for basis comparison plot.}

\item{word_num}{How many GO terms will be used for word cloud plot.}
}
\description{
Perform Gene Ontology (GO) enrichment analysis on each basis to provide biological interpretation
}
\details{
This function annotates the weight matrix \eqn{W}. Each basis of \eqn{W} represents a functional gene cluster. Users can use this annotation to decide which bases should be kept for further analysis.
See details about the GO analysis in \code{clusterProfiler}.
}
\author{
Dongyuan Song
}
