% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisTest.R
\name{basisTest}
\alias{basisTest}
\title{Data-driven basis selection}
\usage{
basisTest(S, X, return_fig = FALSE, adj_method = "BH", mc.cores = 10, nrow = 4)
}
\arguments{
\item{S}{The score matrix of scPNMF.}

\item{X}{The expression matrix of scPNMF.}

\item{return_fig}{If returning the visualization plots. Default is FALSE.}

\item{adj_method}{Multiple test correction method.}

\item{mc.cores}{The number of cores to use.}

\item{nrow}{Rows for facets in plots.}
}
\description{
Perform data-driven basis selection to remove useless bases and keep informative bases. It contains two approaches:
(1) examine bases by correlations with cell library sizes; (2) examine bases by multimodality tests.
}
\details{

}
\author{
Dongyuan Song
}
