% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getApp.R
\name{getInfoGene}
\alias{getInfoGene}
\title{Get informative genes}
\usage{
getInfoGene(W, M = 100, by_basis = FALSE, dim_use = NULL)
}
\arguments{
\item{M}{}

\item{W_s}{}
}
\value{
A vector containing M top informative genes
}
\description{
Get informative genes
}
