% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getApp.R
\name{getProjection}
\alias{getProjection}
\title{Get data projection}
\usage{
getProjection(X, W_sM)
}
\arguments{
\item{X}{The input gene by cell logcount matrix}

\item{W_sM}{The truncated, selected weight matrix}
}
\value{
A K0 * cells matrix representing low-dimensional data projection
}
\description{
Get data projection
}
