% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_selection.R
\name{K_selection}
\alias{K_selection}
\title{Title \eqn{K} selection based on *dev.ortho*}
\usage{
K_selection(
  X,
  K.seq = 10L:30L,
  ncores = 1,
  tol = 0.001,
  maxIter = 500,
  verboseN = FALSE,
  zerotol = 1e-10,
  method = "EucDist",
  label = NULL,
  mu = 1,
  lambda = 0.01,
  seed = 123
)
}
\arguments{
\item{X}{Input data matrix, where rows represent features (genes), columns represent samples (cells). Rownames (gene names) and colnames (cell names) are required.}

\item{K.seq}{An integer vector of candidate \eqn{K}'s. All elements need to be distinct integers, and the total number needs to be no fewer than 5.}

\item{ncores}{The number of cores to use.}

\item{tol}{A threshold below which would be considered converged. Default is 1e-3.}

\item{maxIter}{Number of max iteration times. Default is 500.}

\item{verboseN}{A boolean value indicating whether to print number of iterations.}

\item{zerotol}{A threshold on basis loadings below which would be considered zero. Default is 1e-10.}

\item{method}{A character string indicating which method to be used. One of \code{"EucDist"}, \code{"KL"}, or \code{"DPNMF"}.}

\item{label}{A character vector indicating the cluster type for each cell. Required only when \code{method = "DPNMF"}.}

\item{mu}{A numerical value which controls the penalty term. Larger \code{mu} represents haivier penalization of class distances in \code{DPNMF}. Default is 1.}

\item{lambda}{A numerical value which controls the magnituide of within class distances. Larger \code{lambda} represents larger proportion of within class distances in the total penalty term. Default is 0.01.}

\item{seed}{Random seed of the initialization.}
}
\value{
A list with components:
\describe{
  \item{\code{BestK}}{The selected elbow point \eqn{K}.}
  \item{\code{BestRes}}{The list containing the \code{Weight} and \code{Score} returned by the PNMF algorithm corresponding to \code{BestK}.}
  \item{\code{plot.dev.ortho}}{\code{dev.ortho} plot.}
  \item{\code{AllRes}}{A list containing all PNMF outputs for all candidate \eqn{K}'s in \code{K.seq}.}
}
}
\description{
A low rank \eqn{K} selection method based on *dev.ortho*, the normalized difference between \eqn{\mathbf{W}^T\mathbf{W}} and \eqn{\mathbf{I}}. It runs the PNMF algorithm on every candidate \eqn{K} provided by the user, and find the elbow point \eqn{K} as the suggested one. 

However, this process can be time-consuming as it needs to run the PNMF algorithm on every candidate \eqn{K}. Empirically we recommend the users to skip this section, and directly use \eqn{K=20} in function [PNMFfun()], as we found it reaches stability for most scRNA-seq data.
}
\author{
Kexin Li, \email{aileenlikexin@outlook.com}
}
