% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zheng4}
\alias{zheng4}
\title{A SingelCellExperiment object containing 2,192 genes and 3,994 cells}
\format{
A SingelCellExperiment object with 2,192 rows and 3,994 variables:
}
\source{
Duo, A., Robinson, M. D., and Soneson, C. (2018). A systematic performance evaluation of clustering methods for single-cell rna-seq data. F1000Research, 7.
}
\usage{
zheng4
}
\description{
A SingelCellExperiment object containing 2,192 genes and 3,994 cells
}
\keyword{datasets}
