% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSelect.R
\name{basisSelect}
\alias{basisSelect}
\title{Basis selection function}
\usage{
basisSelect(
  W,
  S = NULL,
  X = NULL,
  toTest = TRUE,
  cor_thres = 0.7,
  pval_thres = 0.01,
  return_fig = FALSE,
  adj_method = "BH",
  mc.cores = 10,
  nrow = 4,
  toAnnotate = FALSE,
  dim_use = NULL
)
}
\arguments{
\item{W}{}

\item{S}{}

\item{X}{}

\item{toTest}{}

\item{cor_thres}{}

\item{pval_thres}{}

\item{return_fig}{}

\item{adj_method}{}

\item{mc.cores}{}

\item{nrow}{}

\item{toAnnotate}{}

\item{dim_use}{}
}
\value{
The selected weight matrix
}
\description{
Basis selection function
}
