% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getApp.R
\name{getInfoGene}
\alias{getInfoGene}
\title{Get informative genes}
\usage{
getInfoGene(W, M = 100, by_basis = FALSE, dim_use = NULL)
}
\arguments{
\item{W}{The weight matrix output by PNMF}

\item{M}{The user-defined informative gene number}

\item{by_basis}{Return informative genes by basis or not}

\item{dim_use}{The bases (columns) to be used in the selected weight matrix, \code{NULL} value uses all bases}
}
\value{
A vector containing M top informative genes
}
\description{
Get informative genes
}
